chore(stella-cli,stella-store): agent delete/assume helpers live in authoring; session_stats is a public module
command_deck.rs is a god file closed to growth; the helpers the AGENTS tab
rewrite added move to command_deck/authoring.rs beside the agent author,
and the store's session rollup is reached as stella_store::session_stats.
The store lib.rs ceiling moves 1729 → 1732 for the module declaration, the
irreducible case AGENTS.md names.

Refs #4334
